Separate logic from homepage to its own file.
